Ajustando classes de localização
