--- ProjetoBancoDeDadosGeoespaciais/src/main/java/dto/LocalizaçaoDTO.java
@@ -1,5 +1,0 @@
-package dto;
-
-public class LocalizaçaoDTO {
-
-}
