--- ProjetoBancoDeDadosGeoespaciais/src/main/java/model/Localizaçao.java
@@ -1,41 +1,0 @@
-package model;
-
-public class Localizaçao {
-	private String estado;
-	private String cidade;
-	private double latitude;
-    private double longitude;
-    
-    public Localizaçao(String estado,String cidade,double latitude,double longitude) {
-    	this.estado=estado;
-    	this.cidade=cidade;
-    	this.latitude=latitude;
-    	this.longitude=longitude;
-    	
-    }
-	public double getLatitude() {
-		return latitude;
-	}
-	public void setLatitude(double latitude) {
-		this.latitude = latitude;
-	}
-	public double getLongitude() {
-		return longitude;
-	}
-	public void setLongitude(double longitude) {
-		this.longitude = longitude;
-	}
-	public String getEstado() {
-		return estado;
-	}
-	public void setEstado(String estado) {
-		this.estado = estado;
-	}
-	public String getCidade() {
-		return cidade;
-	}
-	public void setCidade(String cidade) {
-		this.cidade = cidade;
-	}
-
-}
